requirements.txt
pytest
selenium
pytest-html
pytest-xdist
openpyxl
allure-pytest
pandas
webdriver_manager
python-dotenv